feat(desktop): diff header polish and sidebar co-location
- Open-in-editor chevron wired through electronTrpcClient (bypasses
  the workspaceTrpc provider hijack); deleted files show a toast
  instead of a disabled pill, other failures also toast.
- DiffFileHeader path pill now has border, rounded corners, hover
  state, tooltip (300ms delay) "Open <path>", and a leading
  StatusIndicator showing added/modified/deleted/renamed.
- StatusIndicator promoted to the workspace-level components/ so it's
  shared between the sidebar Changes list and the diff header.
- ChangesFileList broken up per AGENTS.md: FileRow, StatusIndicator,
  and partitionByViewed each live in their own folder. Shared
  ChangedFile type comes from the existing useChangesTab/types.ts.
- useInView co-located under DiffFileEntry (only consumer).
