+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/StatusIndicator/index.ts
@@ -1,0 +1,1 @@
+export { type FileStatus, StatusIndicator } from "./StatusIndicator";
