+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/WorkspaceSidebar/hooks/useChangesTab/components/ChangesFileList/components/FileRow/FileRow.tsx
@@ -1,0 +1,76 @@
+import { Checkbox } from "@superset/ui/checkbox";
+import { memo } from "react";
+import { StatusIndicator } from "renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/StatusIndicator";
+import { FileIcon } from "renderer/screens/main/components/WorkspaceView/RightSidebar/FilesView/utils";
+import type { ChangedFile } from "../../../../types";
+import type { ChangeCategory } from "../../types";
+
+function splitPath(path: string): { dir: string; basename: string } {
+	const lastSlash = path.lastIndexOf("/");
+	if (lastSlash < 0) return { dir: "", basename: path };
+	return {
+		dir: `${path.slice(0, lastSlash)}/`,
+		basename: path.slice(lastSlash + 1),
+	};
+}
+
+interface FileRowProps {
+	file: ChangedFile;
+	category: ChangeCategory;
+	onSelect?: (path: string, category: ChangeCategory) => void;
+	viewed: boolean;
+	onSetViewed: (path: string, next: boolean) => void;
+}
+
+export const FileRow = memo(function FileRow({
+	file,
+	category,
+	onSelect,
+	viewed,
+	onSetViewed,
+}: FileRowProps) {
+	const { dir, basename } = splitPath(file.path);
+
+	return (
+		<div
+			className={`flex w-full items-center gap-1.5 py-1 pr-3 pl-3 text-left text-xs hover:bg-accent/50 ${
+				viewed ? "opacity-60" : ""
+			}`}
+		>
+			<Checkbox
+				checked={viewed}
+				onCheckedChange={(checked) => onSetViewed(file.path, checked === true)}
+				onClick={(e) => e.stopPropagation()}
+				className="size-3.5 shrink-0 border-muted-foreground/50"
+				aria-label={viewed ? "Mark as not viewed" : "Mark as viewed"}
+			/>
+			<button
+				type="button"
+				className="flex min-w-0 flex-1 items-center gap-1.5 text-left"
+				onClick={() => onSelect?.(file.path, category)}
+			>
+				<FileIcon fileName={basename} className="size-3.5 shrink-0" />
+				<span className="flex min-w-0 flex-1 items-baseline overflow-hidden">
+					{dir && <span className="truncate text-muted-foreground">{dir}</span>}
+					<span className="min-w-[120px] truncate font-medium text-foreground">
+						{basename}
+					</span>
+				</span>
+				<span className="ml-auto flex shrink-0 items-center gap-1.5">
+					{(file.additions > 0 || file.deletions > 0) && (
+						<span className="text-[10px] text-muted-foreground">
+							{file.additions > 0 && (
+								<span className="text-green-400">+{file.additions}</span>
+							)}
+							{file.additions > 0 && file.deletions > 0 && " "}
+							{file.deletions > 0 && (
+								<span className="text-red-400">-{file.deletions}</span>
+							)}
+						</span>
+					)}
+					<StatusIndicator status={file.status} />
+				</span>
+			</button>
+		</div>
+	);
+});
