+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/WorkspaceSidebar/hooks/useChangesTab/components/ChangesFileList/types.ts
@@ -1,0 +1,1 @@
+export type ChangeCategory = "against-base" | "staged" | "unstaged";
