+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/WorkspaceSidebar/hooks/useChangesTab/components/ChangesFileList/utils/partitionByViewed/partitionByViewed.ts
@@ -1,0 +1,15 @@
+import type { ChangedFile } from "../../../../types";
+
+export function partitionByViewed(
+	files: ChangedFile[],
+	viewedSet: Set<string>,
+): ChangedFile[] {
+	if (viewedSet.size === 0) return files;
+	const unviewed: ChangedFile[] = [];
+	const viewed: ChangedFile[] = [];
+	for (const file of files) {
+		if (viewedSet.has(file.path)) viewed.push(file);
+		else unviewed.push(file);
+	}
+	return [...unviewed, ...viewed];
+}
